xnu-1228
Imported from https://opensource.apple.com/tarballs/xnu/xnu-1228.tar.gz

+++ bsd/netinet/bootp.h
@@ -1,23 +1,29 @@
 /*
  * Copyright (c) 2000 Apple Computer, Inc. All rights reserved.
  *
- * @APPLE_LICENSE_HEADER_START@
+ * @APPLE_OSREFERENCE_LICENSE_HEADER_START@
  * 
- * The contents of this file constitute Original Code as defined in and
- * are subject to the Apple Public Source License Version 1.1 (the
- * "License").  You may not use this file except in compliance with the
- * License.  Please obtain a copy of the License at
- * http://www.apple.com/publicsource and read it before using this file.
+ * This file contains Original Code and/or Modifications of Original Code
+ * as defined in and that are subject to the Apple Public Source License
+ * Version 2.0 (the 'License'). You may not use this file except in
+ * compliance with the License. The rights granted to you under the License
+ * may not be used to create, or enable the creation or redistribution of,
+ * unlawful or unlicensed copies of an Apple operating system, or to
+ * circumvent, violate, or enable the circumvention or violation of, any
+ * terms of an Apple operating system software license agreement.
  * 
- * This Original Code and all software distributed under the License are
- * distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY KIND, EITHER
+ * Please obtain a copy of the License at
+ * http://www.opensource.apple.com/apsl/ and read it before using this file.
+ * 
+ * The Original Code and all software distributed under the License are
+ * distributed on an 'AS IS' basis, WITHOUT WARRANTY OF ANY KIND, EITHER
  * EXPRESS OR IMPLIED, AND APPLE HEREBY DISCLAIMS ALL SUCH WARRANTIES,
  * INCLUDING WITHOUT LIMITATION, ANY WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE OR NON-INFRINGEMENT.  Please see the
- * License for the specific language governing rights and limitations
- * under the License.
+ * FITNESS FOR A PARTICULAR PURPOSE, QUIET ENJOYMENT OR NON-INFRINGEMENT.
+ * Please see the License for the specific language governing rights and
+ * limitations under the License.
  * 
- * @APPLE_LICENSE_HEADER_END@
+ * @APPLE_OSREFERENCE_LICENSE_HEADER_END@
  */
 /*
  * Bootstrap Protocol (BOOTP).  RFC 951.
@@ -44,7 +50,7 @@ struct bootp {
 	u_char	bp_htype;	/* hardware addr type */
 	u_char	bp_hlen;	/* hardware addr length */
 	u_char	bp_hops;	/* gateway hops */
-	u_long	bp_xid;		/* transaction ID */
+	u_int32_t bp_xid;	/* transaction ID */
 	u_short	bp_secs;	/* seconds since boot began */	
 	u_short	bp_unused;
 	iaddr_t	bp_ciaddr;	/* client IP address */
@@ -68,7 +74,7 @@ struct bootp {
  */
 struct vend {
 	u_char	v_magic[4];	/* magic number */
-	u_long	v_flags;	/* flags/opcodes, etc. */
+	u_int32_t v_flags;	/* flags/opcodes, etc. */
 	u_char	v_unused[56];	/* currently unused */
 };
 #define	VM_STANFORD	"STAN"	/* v_magic for Stanford */
